+++ html/loading.html
@@ -1,0 +1,5 @@
+<div class="loading">
+  <svg class="loading-circle">
+    <circle cx="50%" cy="50%" r="25"></circle>
+  </svg>
+</div>
