+++ html/section-1.html
@@ -1,0 +1,39 @@
+<section class="scroll-section" id="scroll-section-1">
+  <div class="fixed-box">
+    <div class="relative-box">
+      <div class="s1-card-box" id="s1-card-box-0">
+        <div class="absolute-box s1-card" id="s1-card-0">
+          <span class="s1-card-title">올바른 금융.</span>
+          <div class="s1-card-content">
+            <span class="s1-card-desc">
+              올바른 금융을 실현하는 신뢰를 추구하고 모두가 함께 성장하는 사회적
+              가치 실현에 앞서나갑니다.
+            </span>
+          </div>
+        </div>
+      </div>
+      <div class="s1-card-box" id="s1-card-box-1">
+        <div class="absolute-box s1-card" id="s1-card-1">
+          <span class="s1-card-title">혁신의 플랫폼.</span>
+          <div class="s1-card-content">
+            <span class="s1-card-desc">
+              디지털 금융 혁신을 통해 언제, 어디서나, 누구나 필요로하는 금융을
+              제공합니다.
+            </span>
+          </div>
+        </div>
+      </div>
+      <div class="s1-card-box" id="s1-card-box-2">
+        <div class="absolute-box s1-card" id="s1-card-2">
+          <span class="s1-card-title">풍요로운 미래.</span>
+          <div class="s1-card-content">
+            <span class="s1-card-desc">
+              일상 속 특별한 경험을 선사하고, 금융을 넘어 새로운 세상과 풍요로운
+              미래를 연결합니다.
+            </span>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</section>
