+++ html/section-2.html
@@ -1,0 +1,28 @@
+<section class="scroll-section" id="scroll-section-2">
+  <div class="fixed-box">
+    <div class="absolute-box s2-message-box" id="s2-message-box-0">
+      <p id="s2-message-0">
+        <visible class="c-white">하나가</visible>
+        <br class="s2-display-none" />
+        <invisible>새로워진 </invisible>
+        <visible class="c-white">이유</visible>
+      </p>
+    </div>
+    <div class="absolute-box s2-message-box" id="s2-message-box-1">
+      <p id="s2-message-1">
+        <invisible>하나가</invisible>
+        <br class="s2-display-none" />
+        <visible class="s2-gradation-0">새로워진 </visible>
+        <invisible>이유</invisible>
+      </p>
+    </div>
+    <div class="absolute-box s2-message-box" id="s2-message-box-2">
+      <p id="s2-message-2">
+        <invisible>하나가</invisible>
+        <br class="s2-display-none" />
+        <visible class="s2-gradation-1">도전하는 </visible>
+        <invisible>이유</invisible>
+      </p>
+    </div>
+  </div>
+</section>
