+++ html/section-4.html
@@ -1,0 +1,72 @@
+<section class="scroll-section" id="scroll-section-4">
+  <div class="fixed-box">
+    <div class="absolute-box s4-message-box" id="s4-message-0">
+      <p class="s4-message">
+        3분 걸리던
+        <img class="s4-img" src="./assets/images/s4-1.svg" />
+        송금을
+        <br class="s4-display-none" />
+        10초 만에
+      </p>
+    </div>
+    <div class="absolute-box s4-message-box" id="s4-message-1">
+      <p class="s4-message">
+        은행
+        <img class="s4-img" src="./assets/images/s4-2.svg" />
+        점검 시간도
+        <br class="s4-display-none" />
+        기다릴 필요 없이
+      </p>
+    </div>
+    <div class="absolute-box s4-message-box" id="s4-message-2">
+      <p class="s4-message">
+        원할 때 언제든 확인하는
+        <br class="s4-display-none" />
+        <img class="s4-img" src="./assets/images/s4-3.svg" />
+        신용점수
+      </p>
+    </div>
+    <div class="absolute-box s4-message-box" id="s4-message-3">
+      <p class="s4-message">
+        내 모든 금융 내역을
+        <br class="s4-display-none" />
+        <img class="s4-img" src="./assets/images/s4-4.png" />
+        한눈에 조회
+      </p>
+    </div>
+    <div class="absolute-box s4-message-box" id="s4-message-4">
+      <p class="s4-message">
+        가장 좋은
+        <img class="s4-img" src="./assets/images/s4-5.svg" />
+        대출상품
+        <br class="s4-display-none" />
+        발품 팔지 않고 비교
+      </p>
+    </div>
+    <div class="absolute-box s4-message-box" id="s4-message-5">
+      <p class="s4-message">
+        보험금 청구도
+        <br class="s4-display-none" />
+        클릭 몇 번이면 끝
+        <img class="s4-img" src="./assets/images/s4-6.png" />
+      </p>
+    </div>
+    <div class="absolute-box s4-message-box" id="s4-message-6">
+      <p class="s4-message">
+        공공문서
+        <img class="s4-img" src="./assets/images/s4-7.svg" />
+        확인부터
+        <br class="s4-display-none" />
+        납부까지 한 번에
+      </p>
+    </div>
+    <div class="absolute-box s4-message-box" id="s4-message-7">
+      <p class="s4-message">
+        음악 듣듯, 영화 고르듯
+        <br class="s4-display-none" />
+        <img class="s4-img" src="./assets/images/s4-8.svg" />
+        주식 투자를 쉽게
+      </p>
+    </div>
+  </div>
+</section>
