+++ html/section-5.html
@@ -1,0 +1,22 @@
+<section class="scroll-section" id="scroll-section-5">
+  <div class="fixed-box">
+    <div class="s5-video-box">
+      <video muted autoplay loop class="s5-video">
+        <source
+          src="https://static.toss.im/homepage-static/newtoss/the-journey-ending.mp4"
+          type="video/mp4"
+        />
+      </video>
+    </div>
+    <div class="absolute-box s5-message" id="s5-message-0">
+      이제 쉽고 편리한
+      <br class="s5-display-none" />
+      금융을 넘어
+    </div>
+    <div class="absolute-box s5-message c-white" id="s5-message-1">
+      새로운
+      <br class="s5-display-none" />
+      차원으로,
+    </div>
+  </div>
+</section>
