+++ html/section-6.html
@@ -1,0 +1,25 @@
+<section class="scroll-section" id="scroll-section-6">
+  <div class="fixed-box">
+    <div class="relative-box s6-image-box">
+      <img src="./assets/images/hana-symbol.png" />
+    </div>
+    <div class="absolute-box s6-message-box" id="s6-message-box-0">
+      <p id="s6-0">
+        하나가 생각하는 새로운 금융은
+        <br />
+        혜택이 닿지 않던 곳까지 시선을 보내고
+        <br />
+        세심한 변화를 만들어내는 것.
+      </p>
+    </div>
+    <div class="absolute-box s6-message-box" id="s6-message-box-1">
+      <p id="s6-1">
+        모두의 평등한 금융경험을 위해
+        <br />
+        일상에서 금융이 필요한 모든 순간에 귀 기울이고
+        <br />
+        문제를 발견하고, 가장 명쾌한 답을 찾으려 합니다.
+      </p>
+    </div>
+  </div>
+</section>
