+++ html/section-7.html
@@ -1,0 +1,24 @@
+<section class="scroll-section" id="scroll-section-7">
+  <div class="fixed-box">
+    <div class="s7-video-box">
+      <video muted autoplay loop class="s7-video">
+        <source
+          src="https://static.toss.im/homepage-static/newtoss/new-journey-bg.mp4"
+        />
+      </video>
+    </div>
+    <div class="absolute-box s7-message-box" id="s7-message-box-0">
+      <p id="s7-message-0">Growing Together,</p>
+      <p id="s7-message-1">Sharing Happiness</p>
+      <p id="s7-message-2">Original Film by 하나은행</p>
+    </div>
+    <div class="absolute-box s7-message-box" id="s7-message-box-1">
+      <div class="s7-button-box">
+        <button class="s7-button">
+          <img src="./assets/images/s7-1.svg" class="s7-button-img" />
+          <a href="https://youtu.be/u4dzhS_3C2U">오프닝 필름 보기</a>
+        </button>
+      </div>
+    </div>
+  </div>
+</section>
