+++ html/section-8.html
@@ -1,0 +1,28 @@
+<section class="s8-normal-section">
+  <div class="s8-container">
+    <div class="s8-title-box">
+      <p>모두를 위한 <br />하나의 도전은 계속</p>
+    </div>
+    <div class="s8-card-box">
+      <div class="s8-card s8-card-0-bg">
+        <div class="s8-card-title-box">
+          <p>새로운 도전을<br />시작할 준비됐나요?</p>
+        </div>
+        <div class="s8-card-desc-box">
+          <p>나만의 응원카드를 만들어보세요</p>
+        </div>
+        <button class="s8-card-button c-black b-white">
+          <a href="#">응원카드 만들기</a>
+        </button>
+      </div>
+      <div class="s8-card s8-card-1-bg">
+        <div class="s8-card-title-box">
+          <p>지금 진행 중인<br />이벤트는?</p>
+        </div>
+        <button class="s8-card-button c-black b-white">
+          <a href="#">이벤트 살펴보기</a>
+        </button>
+      </div>
+    </div>
+  </div>
+</section>
